fix(weave): dispatch double cast through sqlite null-on-bad-text wrapper
`CastTo` is "double" not "float"; the dispatch set was checking for
"float" so inferred float casts skipped the safer CASE wrapper and fell
back to bare CAST(... AS FLOAT), which returns 0 for non-numeric text
and produced false positives like `bad_rank_text = 0.0`.

+++ weave/trace_server/sqlite_trace_server.py
@@ -5284,7 +5284,7 @@ def _transform_external_calls_field_to_internal_calls_field(
             "json_extract(" + json.dumps(json_column) + ", '" + json_path + "')"
         )
         field = f"CAST({json_extract_sql} AS {sql_type})"
-        if cast in {"int", "float", "bool"}:
+        if cast in {"int", "double", "bool"}:
             json_type_sql = (
                 "json_type(" + json.dumps(json_column) + ", '" + json_path + "')"
             )
